app/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-keycloak==3.7.0
azure-storage-blob==12.19.0
python-jose[cryptography]==3.3.0
requests==2.31.0
gunicorn==21.2.0
werkzeug==3.0.1
configparser==6.0.0
PyJWT==2.8.0
cryptography==41.0.7
